+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [Unreleased]
+
+### Added
+
+- Optionally add SH/SHUI managed classes to shapes graph
+- Optionally import CMEM queryu cayalog to shapes graph
+
+
 ## [3.0.1] 2025-03-13
 
 ### Fixed